requirements.txt
python-telegram-bot
requests
firebase_admin
pydub
ibm_watson